requirements.txt
# Core
openai>=1.12.0
tiktoken>=0.5.0
chromadb>=0.4.0

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# Web Interface
streamlit>=1.31.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.66.0
loguru>=0.7.0

# Development
pytest>=7.4.0
black>=24.0.0
